# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat Atendimento API",
    "version": "1.0.0",
    "description": "\n        API REST para sistema de chat de atendimento em tempo real.\n        \n        ## Funcionalidades\n        - Autenticação JWT\n        - Chat em tempo real com WebSocket\n        - Upload de arquivos (AWS S3)\n        - Gerenciamento de atendentes\n        - Fila de atendimento\n        - Dashboard administrativo\n        - Histórico e análises\n        \n        ## Ambientes\n        - **Desenvolvimento**: http://localhost:5000\n        - **Staging**: http://52.90.17.204:5000\n        - **Produção**: https://api.brsi.com.br\n        \n        ## Autenticação\n        Todas as rotas (exceto login e register) requerem autenticação via JWT.\n        Envie o token no header: Authorization: Bearer {token}\n      ",
    "contact": {
      "name": "BR Sistemas",
      "email": "suporte@brsi.com.br",
      "url": "https://brsi.com.br"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://brsi.com.br/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento (Local)"
    },
    {
      "url": "http://52.90.17.204:5000/api",
      "description": "Servidor de Staging (Testes)"
    },
    {
      "url": "https://api.brsi.com.br/api",
      "description": "Servidor de Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Digite o token JWT obtido no login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload de arquivos",
        "description": "Faz upload de até 5 arquivos para uma conversa. Arquivos são armazenados no AWS S3.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files",
                  "conversationId"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 5,
                    "description": "Arquivos para upload (máximo 5)"
                  },
                  "conversationId": {
                    "type": "string",
                    "description": "ID da conversa associada",
                    "example": "60d0fe4f5311236168a109ca"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "files": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "originalName": {
                            "type": "string"
                          },
                          "mimetype": {
                            "type": "string"
                          },
                          "size": {
                            "type": "number"
                          },
                          "s3Key": {
                            "type": "string"
                          },
                          "s3Bucket": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro no upload ou arquivo inválido"
          },
          "401": {
            "description": "Não autorizado"
          },
          "413": {
            "description": "Arquivo muito grande (limite: 10MB)"
          }
        }
      }
    },
    "/api/files/conversation/{conversationId}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Listar arquivos de uma conversa",
        "description": "Retorna todos os arquivos associados a uma conversa específica",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa",
            "example": "60d0fe4f5311236168a109ca"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de arquivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "filename": {
                        "type": "string"
                      },
                      "originalName": {
                        "type": "string"
                      },
                      "mimetype": {
                        "type": "string"
                      },
                      "size": {
                        "type": "number"
                      },
                      "url": {
                        "type": "string"
                      },
                      "uploadedBy": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/files/download/{fileId}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Baixar arquivo",
        "description": "Retorna URL assinada para download do arquivo do S3",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do arquivo",
            "example": "60d0fe4f5311236168a109cb"
          }
        ],
        "responses": {
          "200": {
            "description": "URL de download gerada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "downloadUrl": {
                      "type": "string",
                      "description": "URL assinada temporária para download (válida por 1 hora)",
                      "example": "https://s3.amazonaws.com/bucket/file?signature=..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Arquivo não encontrado"
          }
        }
      }
    },
    "/api/files/{fileId}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Excluir arquivo",
        "description": "Remove um arquivo do S3 e do banco de dados",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do arquivo",
            "example": "60d0fe4f5311236168a109cb"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Arquivo excluído com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão para excluir este arquivo"
          },
          "404": {
            "description": "Arquivo não encontrado"
          }
        }
      }
    },
...
  }
}